+++ src/main/java/point/by/anton/Point.java
@@ -2,24 +2,24 @@ package point.by.anton;
 
 public class Point {
 
-    private  int x;
-    private  int y;
+    private  double x;
+    private  double y;
 
 
     public Point() {
         this(0,0);
     }
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
